Batch-201/requirements.txt
appdirs==1.4.3
Django==1.8
django-crispy-forms==1.6.1
django-extensions==1.7.9
django-fontawesome==0.3.1
django-registration==2.2
django-tastypie==0.13.3
MySQL-python==1.2.5
mysqlclient==1.3.10
packaging==16.8
psycopg2==2.7.1
PyMySQL==0.7.11
pyparsing==2.2.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
PyYAML==3.12
six==1.10.0
